requirements.txt
gradio>=4.36
python-dotenv>=1.0
smolagents>=0.6
openai>=1.30
litellm>=1.42
pypdf>=4.2
pdf2image>=1.17
pytesseract>=0.3.10
Pillow>=10.3
python-docx>=1.1
openpyxl>=3.1
chardet>=5.2